merge:重构分支开发完成合并到main

+++ logic/system/collider_system.py
@@ -1,40 +1,43 @@
-from logic.interface.system import System
-from logic.context import Context
+from logic.matrix.system import System
+from logic.matrix.context import Context
 from common.logger import logger
 from logic.component.state_component import State
+from logic.matrix.group import Group
+from logic.matrix.matcher import Matcher
 
 
 class ColliderSystem(System):
 
     def __init__(self, context: Context):
         super(ColliderSystem, self).__init__(context)
+        self.box_collider_group: Group = self.context.get_group(Matcher("box2d_collider"))
 
     def update(self):
-        entities = self.context.get_entities()
+        entities = list(self.box_collider_group.entities)
         for i in range(len(entities)):
             entity_a = entities[i]
-            if not entity_a.box_collider or not entity_a.move:
+            if not entity_a.move:
                 continue
 
-            entity_a.box_collider.collider_direction = (0, 0)
+            entity_a.box2d_collider.collider_direction = (0, 0)
             for j in range(len(entities)):
                 if i == j:
                     continue
 
                 entity_b = entities[j]
-                if not entity_b.box_collider:
+                if not entity_b.box2d_collider:
                     continue
-                entity_b.box_collider.collider_direction = (0, 0)
-                if entity_a.box_collider.layer != entity_b.box_collider.layer:
+                entity_b.box2d_collider.collider_direction = (0, 0)
+                if entity_a.box2d_collider.layer != entity_b.box2d_collider.layer:
                     continue
                 if (entity_b.bullet and entity_b.bullet.belong == entity_a.uid) or \
                         (entity_a.bullet and entity_a.bullet.belong == entity_b.uid):
                     continue
 
-                position_a = entity_a.transform.position
-                position_b = entity_b.transform.position
-                distance_x = (entity_a.box_collider.width + entity_b.box_collider.width) / 2
-                distance_y = (entity_a.box_collider.height + entity_b.box_collider.height) / 2
+                position_a = entity_a.get_component("transform").position
+                position_b = entity_b.get_component("transform").position
+                distance_x = (entity_a.box2d_collider.width + entity_b.box2d_collider.width) / 2
+                distance_y = (entity_a.box2d_collider.height + entity_b.box2d_collider.height) / 2
 
                 if position_a[0] >= position_b[0] and position_a[1] >= position_b[1]:
                     delta_x = position_a[0] - position_b[0] - distance_x
@@ -61,20 +64,22 @@ class ColliderSystem(System):
                         continue
                     collider_direction = [delta_x, delta_y]
 
-                if entity_a.transform.rotation == 0.0 or entity_a.transform.rotation == 180.0:
+                if entity_a.get_component("transform").rotation == 0.0 or entity_a.get_component("transform").rotation == 180.0:
                     collider_direction[0] = 0
-                if entity_a.transform.rotation == 90.0 or entity_a.transform.rotation == 270.0:
+                if entity_a.get_component("transform").rotation == 90.0 or entity_a.get_component("transform").rotation == 270.0:
                     collider_direction[1] = 0
 
                 if collider_direction == [0, 0]:
                     continue
 
                 if entity_a.bullet and entity_b.model:
-                    if entity_b.model.model_name in ["wall", "enemy"]:
+                    if entity_b.bullet:
+                        continue
+                    if entity_b.state and entity_b.model.model_name in ["wall", "enemy"]:
                         entity_b.state.state = State.destroy
                     if entity_b.model.model_index in ["symbol"]:
                         entity_b.model.model_index = "destroy"
-                    if entity_b.model.model_name not in ["water"]:
+                    if entity_a.state and entity_b.model.model_name not in ["water"]:
                         entity_a.state.state = State.destroy
                     continue
 
@@ -84,9 +89,9 @@ class ColliderSystem(System):
                         continue
 
                 if entity_a.state.state == State.move:
-                    entity_a.transform.add_position(collider_direction[0], collider_direction[1])
+                    entity_a.get_component("transform").add_position(collider_direction[0], collider_direction[1])
 
-                entity_a.box_collider.collider_direction = (collider_direction[0], collider_direction[1])
+                entity_a.box2d_collider.collider_direction = (collider_direction[0], collider_direction[1])
                 logger.debug(f"box collider detect {collider_direction}: {entity_a.uid} -> {entity_b.uid}")
